refactor: 채팅 소켓 스토어를 메세지 엔티티로 이동 (#99)
