+++ src/entities/message/api/socket.ts
